requirements.txt
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
deemix==3.6.6
deezer-py==1.3.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mutagen==1.47.0
nest-asyncio==1.6.0
pycryptodomex==3.23.0
python-telegram-bot==22.3
requests==2.32.5
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
